Refactor to expose the MultibodyTree model and avoid cyclic dependency on the plant
